Switching to parameter removeScatteredOutliers
This parameter is a boolean, which makes it less confusing

Now, outliers are removed at the beginning of the second
outer iteration of smooth fitting instead of having an
outer iteration of their own.  The behavior should not
be qualitatively different and the results of all vortex
tests seem to be similar to before the change.

I changed all the defaultParameters.ascii to use the new parameter.
I also added a missing script (writeParameterFiles.csh) and fixed
or added a few other scripts that had gotten lost or damaged in
some previous commit.